refactor: VDR e2e tests
